build small bootable ISO with standalone UEFI shell
Although OVMF auto-generates a UEFI boot option for the memory-mapped,
firmware image-contained UEFI shell in:

  PlatformBdsPolicyBehavior()      [OvmfPkg/Library/PlatformBdsLib/BdsPlatform.c]
    BdsLibEnumerateAllBootOption() [IntelFrameworkModulePkg/Library/GenericBdsLib/BdsBoot.c]

that boot option is unexpressible with the OpenFirmware device paths that
qemu generates for the -boot XXX and -device YYY,bootindex=Z options.
Hence the memory-mapped UEFI shell cannot be selected for booting in the
libvirt domain XML or on the qemu command line. It is selectable only by
entering the OVMF setup TUI at the TianoCore splash screen.

Build a minimal ISO image that is bootable by OVMF, and copy the UEFI
shell binary to the default boot loader location on it. The management
layer can request OVMF to boot a CD-ROM (IDE or virtio-scsi) just fine.

This is a port of the
<https://bugzilla.redhat.com/show_bug.cgi?id=1147592> patch to Gerd's
public RPMs.

Suggested-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Laszlo Ersek <lersek@redhat.com>