support building OVMF with distcc
Due to the way the ARM cross-compilers are configured, they already
utilize distcc when the latter is set up in the build environment. Enable
the same for OVMF.

Note that we're explicitly relying on the %_smp_mflags / %jobs RPM macros.
Distcc users should either add the following to their $HOME/.rpmmacros:

  %_smp_mflags -j%(distcc -j)

or pass it on the command line as:

  rpmbuild --define "_smp_mflags -j$(distcc -j)" ...

Signed-off-by: Laszlo Ersek <lersek@redhat.com>